% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2.R
\name{draw_plot}
\alias{draw_plot}
\title{Draw the plot of daily cases and deaths numbers for locations.}
\usage{
draw_plot(data, x, y, fill)
}
\arguments{
\item{data}{The dataset, here we are using "owid_eu_long".}

\item{x}{"location" as the x-axis of this graph.}

\item{y}{"count" as the y-axis of this graph, representing numbers of daily cases and deaths.}

\item{fill}{"type_new"}
}
\description{
Draw the plot of daily cases and deaths numbers for locations.
}
