% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eutable_draft.R
\docType{data}
\name{eutable_draft}
\alias{eutable_draft}
\title{Health and Coronavirus oriented information in European Union (EU)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 236 rows and 7 columns.
}
\source{
\href{https://ourworldindata.org/coronavirus-source-data}{Our World in Data}
}
\usage{
eutable_draft
}
\description{
With the dimension of 236 x 7, this dataset exhibits the cumulative cases and deaths numbers on the 5th of each month of 27 EU member countries. Population and a significant health indicator, life expectancy, are also displayed. Additionally, case fatality rate is computed by \verb{Total cases} divided by \verb{Total deaths}.
}
\keyword{datasets}
