% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r1.R
\name{ic}
\alias{ic}
\title{Generate the format for tag$style(HTML())}
\usage{
ic(id, color)
}
\arguments{
\item{id}{The id for each individual \code{tag$style(HTML())}.}

\item{color}{The color of the output.}
}
\description{
Generate the format for tag$style(HTML())
}
