% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owid.R
\docType{data}
\name{owid}
\alias{owid}
\title{Global condition of coronavirus-related information}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 48169 rows and 41 columns.
}
\source{
Our World in Data
}
\usage{
owid
}
\description{
Consisted of 41 variables, this dataset documents various coronavirus-related information, such as total deaths, new cases and life expectancy etc., over 211 countries.
}
\keyword{datasets}
