% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owid_eu.R
\docType{data}
\name{owid_eu}
\alias{owid_eu}
\title{Condition of Coronavirus-related information in European Union (EU)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6712 rows and 8 columns.
}
\source{
Our World in Data
}
\usage{
owid_eu
}
\description{
This dataset is transformed from \code{covideu::owid}. Originally there are 211 countries in \code{covideu::owid}, however, 27 EU member countries are selected to examine the impact of Covid-19 among the EU. 41 variables are narrowed down to 8 through \code{filter()} function.
}
\keyword{datasets}
