% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owid_eu_long.R
\docType{data}
\name{owid_eu_long}
\alias{owid_eu_long}
\title{Condition of Coronavirus-related information in European Union (EU) (Long form)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 13424 rows and 8 columns.
}
\source{
\href{https://ourworldindata.org/coronavirus-source-data}{Our World in Data}
}
\usage{
owid_eu_long
}
\description{
Comprised with 8 variables in relevance with Coronavirus, this dataset share the same information as \code{owid_eu} dataset. The major difference is that this dataset presents \code{new_cases} and \code{new_deaths} in a long format, so that daily cases and deaths numbers can be calculated respectively and can be found in \code{count} column.
}
\keyword{datasets}
